# src/energy_optimizer/openapi-docs.yml
openapi: 3.1.0
info:
  title: Energy Optimizer
  version: 0.1.0
paths:
...
  /api/v1/heat-pump:
    get:
      summary: Persisted Heat Pump
      description: Return the latest imported heat-pump snapshot with polling freshness.
      operationId: persisted_heat_pump_api_v1_heat_pump_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeatPumpLoadResponse'
        '404':
          description: No imported heat-pump snapshot
        '503':
          description: Mappings or persistence unavailable
    post:
      summary: Heat Pump
      description: Validate normalized heat-pump load and electrical flexibility constraints.
      operationId: heat_pump_api_v1_heat_pump_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeatPumpLoad'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeatPumpLoadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/battery:
    post:
      summary: Battery
      description: Validate a versioned hourly battery state and capabilities object.
      operationId: battery_api_v1_battery_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatteryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatteryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/electricity-prices:
    post:
      summary: Electricity Prices
      description: Validate normalized hourly electricity-price data.
      operationId: electricity_prices_api_v1_electricity_prices_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElectricityPriceRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElectricityPriceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/household-load:
    get:
      summary: Persisted Household Load
      description: Return the latest persisted normalized household-load provider
        data.
      operationId: persisted_household_load_api_v1_household_load_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdLoadResponse'
        '404':
          description: No persisted provider data is available
        '503':
          description: Provider data persistence is unavailable
    post:
      summary: Household Load
      description: Validate a versioned hourly household-load data series.
      operationId: household_load_api_v1_household_load_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdLoadRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdLoadResponse'
        '503':
          description: Provider data could not be persisted
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/historic/household-load:
    get:
      summary: Historic Household Load
      description: Return persisted household-load actuals for a requested time range.
      operationId: historic_household_load_api_v1_historic_household_load_get
      parameters:
      - name: start_time
        in: query
        required: true
        schema:
          type: string
          format: date-time
          description: Inclusive timezone-aware range start
          title: Start Time
        description: Inclusive timezone-aware range start
      - name: end_time
        in: query
        required: true
        schema:
          type: string
          format: date-time
          description: Exclusive timezone-aware range end
          title: End Time
        description: Exclusive timezone-aware range end
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoricHouseholdLoadResponse'
        '404':
          description: No persisted household-load data is available
        '503':
          description: Persisted household-load data is unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/pv-generation:
    post:
      summary: Pv Generation
      description: Validate a versioned hourly PV-generation data series.
      operationId: pv_generation_api_v1_pv_generation_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PvGenerationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PvGenerationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/grid-flow:
    get:
      summary: Persisted Grid Flow
      description: Return the latest persisted normalized grid-flow provider data.
      operationId: persisted_grid_flow_api_v1_grid_flow_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GridFlowResponse'
        '404':
          description: No persisted provider data is available
        '503':
          description: Provider data persistence is unavailable
    post:
      summary: Grid Flow
      description: Validate a versioned hourly grid import and export data series.
      operationId: grid_flow_api_v1_grid_flow_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GridFlowRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GridFlowResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    BatteryRequest:
      properties:
        schema_version:
          type: string
          const: '1'
          title: Schema Version
          description: Version of this API contract
        start_time:
          type: string
          format: date-time
          title: Start Time
          description: Timezone-aware start of the series
        interval_minutes:
          type: integer
          const: 60
          title: Interval Minutes
          description: Duration of every series interval; hourly data requires 60
        state_of_charge_kwh:
          items:
            type: number
            maximum: 100000.0
            minimum: 0.0
          type: array
          maxItems: 87672
          minItems: 1
          title: State Of Charge Kwh
          description: Battery state of charge in kWh, one value per interval
        capacity_kwh:
          type: number
          maximum: 100000.0
          exclusiveMinimum: 0.0
          title: Capacity Kwh
          description: Usable battery capacity in kWh
        minimum_soc_kwh:
          type: number
          maximum: 100000.0
          minimum: 0.0
          title: Minimum Soc Kwh
          description: Minimum allowed state of charge in kWh
        maximum_soc_kwh:
          type: number
          maximum: 100000.0
          exclusiveMinimum: 0.0
          title: Maximum Soc Kwh
          description: Maximum allowed state of charge in kWh
        initial_soc_kwh:
          type: number
          maximum: 100000.0
          minimum: 0.0
          title: Initial Soc Kwh
          description: State of charge before the first interval in kWh
        maximum_charge_kw:
          type: number
          maximum: 100000.0
          exclusiveMinimum: 0.0
          title: Maximum Charge Kw
          description: Maximum charging power in kW
        maximum_discharge_kw:
          type: number
          maximum: 100000.0
          exclusiveMinimum: 0.0
          title: Maximum Discharge Kw
          description: Maximum discharging power in kW
        battery_efficiency:
          type: number
          maximum: 1.0
          exclusiveMinimum: 0.0
          title: Battery Efficiency
          description: Measured battery round-trip efficiency
        unit:
          type: string
          const: kWh
          title: Unit
          description: Unit used by energy fields
        power_unit:
          type: string
          const: kW
          title: Power Unit
          description: Unit used by power fields
        source:
          anyOf:
          - $ref: '#/components/schemas/SourceMetadata'
          - type: 'null'
          description: Optional source metadata for externally supplied data
      additionalProperties: false
      type: object
      required:
      - schema_version
      - start_time
      - interval_minutes
      - state_of_charge_kwh
      - capacity_kwh
      - minimum_soc_kwh
      - maximum_soc_kwh
      - initial_soc_kwh
      - maximum_charge_kw
      - maximum_discharge_kw
      - battery_efficiency
      - unit
      - power_unit
      title: BatteryRequest
      description: Versioned battery state and capability data at the API boundary.
    BatteryResponse:
      properties:
        status:
          type: string
          const: validated
          title: Status
        schema_version:
          type: string
          const: '1'
          title: Schema Version
        start_time:
          type: string
          format: date-time
          title: Start Time
        interval_minutes:
          type: integer
          const: 60
          title: Interval Minutes
        state_of_charge_kwh:
          items:
            type: number
          type: array
          title: State Of Charge Kwh
        capacity_kwh:
          type: number
          title: Capacity Kwh
        minimum_soc_kwh:
          type: number
          title: Minimum Soc Kwh
        maximum_soc_kwh:
          type: number
          title: Maximum Soc Kwh
        initial_soc_kwh:
          type: number
          title: Initial Soc Kwh
        maximum_charge_kw:
          type: number
          title: Maximum Charge Kw
        maximum_discharge_kw:
          type: number
          title: Maximum Discharge Kw
        battery_efficiency:
          type: number
          title: Battery Efficiency
        unit:
          type: string
          const: kWh
          title: Unit
        power_unit:
          type: string
          const: kW
          title: Power Unit
        source:
          anyOf:
          - $ref: '#/components/schemas/SourceMetadata'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - status
      - schema_version
      - start_time
      - interval_minutes
      - state_of_charge_kwh
      - capacity_kwh
      - minimum_soc_kwh
      - maximum_soc_kwh
      - initial_soc_kwh
      - maximum_charge_kw
      - maximum_discharge_kw
      - battery_efficiency
      - unit
      - power_unit
      title: BatteryResponse
      description: Response returned after battery data passes validation.
...
    ElectricityPriceRequest:
      properties:
        schema_version:
          type: string
          const: '1'
          title: Schema Version
          description: Version of this API contract
        timestamps:
          items:
            type: string
            format: date-time
          type: array
          maxItems: 87672
          minItems: 1
          title: Timestamps
          description: Timezone-aware hourly timestamps in ascending order
        interval_minutes:
          type: integer
          const: 60
          title: Interval Minutes
          description: Duration of every price interval; hourly data requires 60
        import_price_eur_per_kwh:
          items:
            type: number
            maximum: 100.0
            minimum: -100.0
          type: array
          maxItems: 87672
          minItems: 1
          title: Import Price Eur Per Kwh
          description: Grid import price in EUR/kWh, one value per timestamp
        export_price_eur_per_kwh:
          items:
            type: number
            maximum: 100.0
            minimum: -100.0
          type: array
          maxItems: 87672
          minItems: 1
          title: Export Price Eur Per Kwh
          description: Grid export price in EUR/kWh, one value per timestamp
        unit:
          type: string
          const: EUR/kWh
          title: Unit
          description: Unit used by price fields
        source:
          $ref: '#/components/schemas/SourceMetadata'
          description: Provider-independent source metadata
        retrieved_at:
          type: string
          format: date-time
          title: Retrieved At
          description: Time when the source data was retrieved
        expires_at:
          type: string
          format: date-time
          title: Expires At
          description: Time after which the data is stale
      additionalProperties: false
      type: object
      required:
      - schema_version
      - timestamps
      - interval_minutes
      - import_price_eur_per_kwh
      - export_price_eur_per_kwh
      - unit
      - source
      - retrieved_at
      - expires_at
      title: ElectricityPriceRequest
      description: Versioned normalized hourly electricity-price data.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HeatPumpLoad:
      properties:
        schema_version:
          type: string
          const: '1'
          title: Schema Version
        start_time:
          type: string
          format: date-time
          title: Start Time
        interval_minutes:
          type: integer
          const: 60
          title: Interval Minutes
        load_kw:
          items:
            type: number
            maximum: 1000.0
            minimum: 0.0
          type: array
          maxItems: 168
          minItems: 1
          title: Load Kw
        available:
          items:
            type: boolean
          type: array
          maxItems: 168
          minItems: 1
          title: Available
        minimum_power_kw:
          type: number
          maximum: 1000.0
          minimum: 0.0
          title: Minimum Power Kw
        maximum_power_kw:
          type: number
          maximum: 1000.0
          exclusiveMinimum: 0.0
          title: Maximum Power Kw
        required_energy_kwh:
          type: number
          maximum: 168000.0
          minimum: 0.0
          title: Required Energy Kwh
        unit:
          type: string
          const: kW
          title: Unit
        energy_unit:
          type: string
          const: kWh
          title: Energy Unit
        source:
          anyOf:
          - $ref: '#/components/schemas/HeatPumpSource'
          - type: 'null'
        retrieved_at:
          type: string
          format: date-time
          title: Retrieved At
        latest_observation_at:
          type: string
          format: date-time
          title: Latest Observation At
      additionalProperties: false
      type: object
      required:
      - schema_version
      - start_time
      - interval_minutes
      - load_kw
      - available
      - minimum_power_kw
      - maximum_power_kw
      - required_energy_kwh
      - unit
      - energy_unit
      - retrieved_at
      - latest_observation_at
      title: HeatPumpLoad
      description: An exact electrical energy requirement movable across available
        hours.
    HeatPumpLoadResponse:
      properties:
        schema_version:
          type: string
          const: '1'
          title: Schema Version
        start_time:
          type: string
          format: date-time
          title: Start Time
        interval_minutes:
          type: integer
          const: 60
          title: Interval Minutes
        load_kw:
          items:
            type: number
            maximum: 1000.0
            minimum: 0.0
          type: array
          maxItems: 168
          minItems: 1
          title: Load Kw
        available:
          items:
            type: boolean
          type: array
          maxItems: 168
          minItems: 1
          title: Available
        minimum_power_kw:
          type: number
          maximum: 1000.0
          minimum: 0.0
          title: Minimum Power Kw
        maximum_power_kw:
          type: number
          maximum: 1000.0
          exclusiveMinimum: 0.0
          title: Maximum Power Kw
        required_energy_kwh:
          type: number
          maximum: 168000.0
          minimum: 0.0
          title: Required Energy Kwh
        unit:
          type: string
          const: kW
          title: Unit
        energy_unit:
          type: string
          const: kWh
          title: Energy Unit
        source:
          anyOf:
          - $ref: '#/components/schemas/HeatPumpSource'
          - type: 'null'
        retrieved_at:
          type: string
          format: date-time
          title: Retrieved At
        latest_observation_at:
          type: string
          format: date-time
          title: Latest Observation At
        status:
          type: string
          enum:
          - validated
          - stale
          title: Status
          default: validated
        freshness:
          type: string
          enum:
          - fresh
          - stale
          - unknown
          title: Freshness
          default: unknown
      additionalProperties: false
      type: object
      required:
      - schema_version
      - start_time
      - interval_minutes
      - load_kw
      - available
      - minimum_power_kw
      - maximum_power_kw
      - required_energy_kwh
      - unit
      - energy_unit
      - retrieved_at
      - latest_observation_at
      title: HeatPumpLoadResponse
    HeatPumpSource:
      properties:
        provider:
          type: string
          maxLength: 100
          minLength: 1
          title: Provider
        entity_id:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Entity Id
      additionalProperties: false
      type: object
      required:
      - provider
      title: HeatPumpSource
...